test(prevention_policy): use export_test.go for precedence filter tests
Move the filterPoliciesByCID, distinctCIDs, and stripChecksum unit tests
into the existing external precedence test file, exposing the helpers and
policyRef through export_test.go instead of a separate internal test file.
