+++ internal/prevention_policy/export_test.go
@@ -3,9 +3,19 @@ package preventionpolicy
 type (
 	PreventionPoliciesDataSource      = preventionPoliciesDataSource
 	PreventionPoliciesDataSourceModel = preventionPoliciesDataSourceModel
+	PolicyRef                         = policyRef
 )
 
 var (
 	FilterPoliciesByIDs        = filterPoliciesByIDs
 	FilterPoliciesByAttributes = filterPoliciesByAttributes
+
+	FilterPoliciesByCID = filterPoliciesByCID
+	DistinctCIDs        = distinctCIDs
+	StripChecksum       = stripChecksum
 )
+
+// NewPolicyRef builds a policyRef for tests in the external test package.
+func NewPolicyRef(id, cid, name string) policyRef {
+	return policyRef{id: id, cid: cid, name: name}
+}
